Some basic cleanup to get things out of app.js
